respaldando modificaciones

+++ app/Http/Controllers/PrePostulacionUniversidadController.php
@@ -8,6 +8,7 @@ use App\PreOtroFinanciamiento;
 use App\Postulante;
 use App\Continente;
 use App\PostPostulacionUniversidad;
+use App\PostOtroFinanciamiento;
 use App\PrePostulacionUniversidad;
 use Illuminate\Http\Request;
 
@@ -83,8 +84,32 @@ class PrePostulacionUniversidadController extends Controller {
 			$postPostulacion = PostPostulacionUniversidad::where('postulante',$postulante->id)->first();
 			if($postPostulacion)
 			{
+				$parametros['id'] = $postPostulacion->id;
+				$parametros['programa'] = $postPostulacion->tipo;
+				$parametros['instituto'] = $postPostulacion->instituto;
+				$parametros['nomLab'] = $postPostulacion->laboratorio;
+				$parametros['contacto'] = $postPostulacion->contacto_uach;
+				$parametros['area'] = $postPostulacion->area;
+				$parametros['nombreP'] = $postPostulacion->nombre_maestria;
+				$parametros['postulante'] = $postPostulacion->postulante;
+				$parametros['anio'] = $postPostulacion->anio;
+				$parametros['semestre'] = $postPostulacion->duracion;
+				$parametros['financiamiento'] = $postPostulacion->financiamiento;
+				$parametros['financiamiento_nombre'] = $postPostulacion->financiamientoR->nombre;
+				$parametros['desde'] = $postPostulacion->desde;
+				$parametros['hasta'] = $postPostulacion->hasta;
+				$parametros['facultad'] = $postPostulacion->facultad;
+				$parametros['campus_sede'] = $postPostulacion->facultadR->campusSedesR->id;
+				$parametros['pais'] = $postPostulacion->facultadR->campusSedesR->ciudadR->paisR->id;
+				$parametros['continente'] = $postPostulacion->facultadR->campusSedesR->ciudadR->paisR->continente;
 
-dd($postPostulacion->toArray());
+				if($postPostulacion->financiamiento == 4 or $postPostulacion->financiamiento == 5){
+					$otroFinanciamiento = PostOtroFinanciamiento::find($postPostulacion->postulante);
+					$parametros['descripcion'] = $otroFinanciamiento->descripcion;
+
+			
+				}
+				return view('postulacion.postulacion_universidad.edit',compact('continentes','parametros'));
 
 			}
 			else
@@ -116,40 +141,87 @@ dd($postPostulacion->toArray());
 				]);
 		}
 
+
+		else{
+
+
+			$postPostulacion = new PostPostulacionUniversidad();
+
+			$postPostulacion->postulante = $postulante->id;
+			$postPostulacion->tipo = $request->get('programa');
+			$postPostulacion->anio = $request->get('anio');
+			$postPostulacion->duracion = $request->get('semestre');
+			$postPostulacion->desde = $request->get('desde');
+			$postPostulacion->hasta = $request->get('hasta');
+			$postPostulacion->area = $request->get('area');
+			$postPostulacion->nombre_maestria = $request->get('nombreP');
+			$postPostulacion->laboratorio = $request->get('nomLab');
+			$postPostulacion->contacto_uach = $request->get('contacto');
+			$postPostulacion->instituto = $request->get('instituto');
+			$postPostulacion->facultad = $request->get('facultad');
+			$postPostulacion->financiamiento = $request->get('financiamiento');
+			if ($request->get('financiamiento') ==='4' or $request->get('financiamiento') ==='5')
+			{
+			   $otroFinanciamiento = new PostOtroFinanciamiento();
+			   $otroFinanciamiento->descripcion = $request->get('descripcion');
+			   $otroFinanciamiento->postulante = $postPostulacion->postulante;
+			   $otroFinanciamiento->save();
+			}
+
+			$postPostulacion->save();
+		}
+
 		return response()->json([
-				'message'=> 'se Guardó la universidad Correctamente'
+				'message'=> 'Se guardaron correctamente sus adtos de intercambio'
 				]);
 	}
 
 	public function putUpdate(PrePostulacionUniversidadRequest $request,Guard $auth){
 
-		$prePostulacion = PrePostulacionUniversidad::find($request->get('id'));
+		$postulante  = Postulante::where('user_id',$auth->id())->first();
 
-		//dd();
-		if($prePostulacion->preOtroFinanciamientosR->count()){
-			if ($request->get('financiamiento') != '4'){
 
-				if($request->get('financiamiento') != '5'){
 
-						$otroFinanciamiento = PreOtroFinanciamiento::find($prePostulacion->id);
-						$otroFinanciamiento->delete();
 
-				}
-			} 
+		if($postulante->tipo_estudio === "Pregrado"){
+
+			$prePostulacion = PrePostulacionUniversidad::find($request->get('id'));
+
+			//dd();
+			if($prePostulacion->preOtroFinanciamientosR->count()){
+				if ($request->get('financiamiento') != '4'){
+
+					if($request->get('financiamiento') != '5'){
+
+							$otroFinanciamiento = PreOtroFinanciamiento::find($prePostulacion->id);
+							$otroFinanciamiento->delete();
+
+					}
+				} 
+			}
+			$prePostulacion->fill($request->all());
+			$prePostulacion->save();
+			if ($request->get('financiamiento') ==='4' or $request->get('financiamiento') ==='5')
+			{
+				$otroFinanciamiento = PreOtroFinanciamiento::firstOrNew(array('pre_postulacion_universidad'=> $prePostulacion->id));
+				//dd($otroFinanciamiento);
+			   	$otroFinanciamiento->descripcion = $request->get('descripcion');
+			   	$otroFinanciamiento->save();
+			}
+
+			return response()->json([
+					'message'=> 'Se han actualizado los datos de la pestaña referente a intercambio'
+					]);
+
+
+			
 		}
-		$prePostulacion->fill($request->all());
-		$prePostulacion->save();
-		if ($request->get('financiamiento') ==='4' or $request->get('financiamiento') ==='5')
-		{
-			$otroFinanciamiento = PreOtroFinanciamiento::firstOrNew(array('pre_postulacion_universidad'=> $prePostulacion->id));
-			//dd($otroFinanciamiento);
-		   	$otroFinanciamiento->descripcion = $request->get('descripcion');
-		   	$otroFinanciamiento->save();
-		}
 
-		return response()->json([
-				'message'=> 'Se han actualizado los datos de la pestaña referente a intercambio'
-				]);
+
+		else{
+
+			dd("jojo");
+		}
 
 	}
